TFloatRect is now using double
Former-commit-id: 0ba8a7f6d92fe47c90f0494451564448498005f2

+++ src/model/IPosition.cpp
@@ -111,7 +111,7 @@ TFloatRect IPosition::getBoundingRect() const
 	}
 
 	//Scale
-	result.setSize( result.size() * getScale() );
+	result.setSize( result.size() * (double)getScale() );
 	//Pos: rectangle is centered on getPos().
 	result.setPos( getPos() - TFloatPoint( 0.5f * result.width() , 0.5f * result.height() ) );
 	return result;
